took care of the refactoring , can u please check the changes .
