+++ listenbrainz/webserver/views/export.py
@@ -7,6 +7,7 @@ from psycopg2 import DatabaseError
 from sqlalchemy import text
 
 from listenbrainz.garage import get_error_code, get_garage_client, get_user_data_export_bucket
+from listenbrainz.db import user_data_export
 from listenbrainz.webserver import db_conn
 from listenbrainz.webserver.decorators import web_listenstore_needed
 from listenbrainz.webserver.errors import APIInternalServerError, APINotFound, APIBadRequest
@@ -21,40 +22,10 @@ export_bp = Blueprint("export", __name__)
 def create_export_task():
     """ Add a request to export the user data to an archive in background. """
     try:
-        query = """
-            INSERT INTO user_data_export (user_id, type, status, progress)
-                 VALUES (:user_id, :type, 'waiting', :progress)
-            ON CONFLICT (user_id, type)
-                  WHERE status = 'waiting' OR status = 'in_progress'
-             DO NOTHING
-              RETURNING id, type, available_until, created, progress, status, filename
-        """
-        result = db_conn.execute(text(query), {
-            "user_id": current_user.id,
-            "type": "export_all_user_data",
-            "progress": "Your data export will start soon."
-        })
-        export = result.first()
-
-        if export is not None:
-            query = "INSERT INTO background_tasks (user_id, task, metadata) VALUES (:user_id, :task, :metadata) ON CONFLICT DO NOTHING RETURNING id"
-            result = db_conn.execute(text(query), {
-                "user_id": current_user.id,
-                "task": "export_all_user_data",
-                "metadata": json.dumps({"export_id": export.id})
-            })
-            task = result.first()
-            if task is not None:
-                db_conn.commit()
-                return jsonify({
-                    "export_id": export.id,
-                    "type": export.type,
-                    "available_until": export.available_until.isoformat() if export.available_until is not None else None,
-                    "created": export.created.isoformat(),
-                    "progress": export.progress,
-                    "status": export.status,
-                    "filename": export.filename,
-                })
+        export_data = user_data_export.request_user_data_export(db_conn, current_user.id)
+        if export_data is not None:
+            db_conn.commit()
+            return jsonify(export_data)
 
         # task already exists in queue, rollback new entry
         db_conn.rollback()
@@ -70,22 +41,10 @@ def create_export_task():
 @web_listenstore_needed
 def get_export_task(export_id):
     """ Retrieve the requested export's data if it belongs to the specified user """
-    result = db_conn.execute(
-        text("SELECT * FROM user_data_export WHERE user_id = :user_id AND id = :export_id"),
-        {"user_id": current_user.id, "export_id": export_id}
-    )
-    row = result.first()
-    if row is None:
+    export_data = user_data_export.get_export_task(db_conn, current_user.id, export_id)
+    if export_data is None:
         raise APINotFound("Export not found")
-    return jsonify({
-        "export_id": row.id,
-        "type": row.type,
-        "available_until": row.available_until.isoformat() if row.available_until is not None else None,
-        "created": row.created.isoformat(),
-        "progress": row.progress,
-        "status": row.status,
-        "filename": row.filename,
-    })
+    return jsonify(export_data)
 
 
 @export_bp.get("/list/")
@@ -93,20 +52,8 @@ def get_export_task(export_id):
 @web_listenstore_needed
 def list_export_tasks():
     """ Retrieve the all export tasks for the current user """
-    result = db_conn.execute(
-        text("SELECT * FROM user_data_export WHERE user_id = :user_id ORDER BY created DESC"),
-        {"user_id": current_user.id}
-    )
-    rows = result.mappings().all()
-    return jsonify([{
-        "export_id": row.id,
-        "type": row.type,
-        "available_until": row.available_until.isoformat() if row.available_until is not None else None,
-        "created": row.created.isoformat(),
-        "progress": row.progress,
-        "status": row.status,
-        "filename": row.filename,
-    } for row in rows])
+    export_tasks = user_data_export.list_export_tasks(db_conn, current_user.id)
+    return jsonify(export_tasks)
 
 
 @export_bp.post("/download/<export_id>/")
@@ -151,16 +98,8 @@ def download_export_archive(export_id):
 @web_listenstore_needed
 def delete_export_archive(export_id):
     """ Delete the specified export archive """
-    result = db_conn.execute(
-        text("DELETE FROM user_data_export WHERE user_id = :user_id AND id = :export_id RETURNING filename"),
-        {"user_id": current_user.id, "export_id": export_id}
-    )
-    row = result.first()
-    if row is not None:
-        db_conn.execute(
-            text("DELETE FROM background_tasks WHERE user_id = :user_id AND (metadata->'export_id')::int = :export_id"),
-            {"user_id": current_user.id, "export_id": export_id}
-        )
+    success = user_data_export.delete_export_task(db_conn, current_user.id, export_id)
+    if success:
         db_conn.commit()
         # archive is deleted from garage by cronjob
         return jsonify({"success": True})
